Add files via upload
Hotel Management is a database management based Project which use Mysql for managing the hotel related data. the project makes use of HTML, CSS for frontend. JDBC is used to connect database and java. Tomcat server is used to run the frontend part .Here for backend JSP is used .
Platform: Eclipse EE version